% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{counterparts}
\alias{counterparts}
\title{World Bank Counterparts}
\format{
A data frame with the following columns:
\describe{
\item{counterpart_id}{Unique identifier for the counterpart}
\item{counterpart_name}{Name of the counterpart entity}
\item{geography_iso3}{ISO 3166-1 alpha-3 code, where available}
\item{geography_iso2}{ISO 3166-1 alpha-2 code, where available}
\item{geography_type}{Type of geographic entity (e.g., "Country",
"Region", "Institution")}
\item{counterpart_type}{Classification of the counterpart (e.g.,
"Country", "Global MDBs", "Global IFIs")}
}
}
\source{
World Bank International Debt Statistics API
}
\usage{
counterparts
}
\description{
A dataset containing counterpart entities from the World Bank's IDS API.
}
\keyword{datasets}
