% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{country_aggregates}
\alias{country_aggregates}
\title{Standardized Country and Aggregate Mappings to IMF and World Bank Areas}
\format{
A data frame with 2587 rows and 3 columns:
\describe{
\item{country_name}{Standardized country name}
\item{country_group}{Standardized aggregate name}
\item{group_type}{Either "IMF" or "World Bank"}
}
}
\source{
\url{https://github.com/t-emery/sais-susfin_data/blob/main/datasets/
imf_wb_country_groups.csv}
}
\usage{
country_aggregates
}
\description{
A dataset containing geographic groupings mapped to area names used by the
IMF and World Bank.
}
\keyword{datasets}
