% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{country_codes}
\alias{country_codes}
\title{Country Data with ISO Codes and World Bank Information}
\format{
A data frame with columns:
\describe{
\item{country}{Country name}
\item{is_wb_country}{Logical indicating if the country is recognized by the
World Bank}
\item{wb_name}{World Bank name for the country}
\item{cldr_short_en}{CLDR short name for the country}
\item{iso2c}{ISO 3166-1 alpha-2 code}
\item{iso3c}{ISO 3166-1 alpha-3 code}
\item{iso3n}{ISO 3166-1 numeric code}
\item{imf}{IMF code}
\item{continent}{Continent name}
\item{country_name_en_regex}{Standardized country name}
}
}
\source{
\url{https://github.com/t-emery/wbhelpr/blob/master/data-raw/
wb_countrycode.csv}
}
\usage{
country_codes
}
\description{
A dataset containing country information, including ISO codes and World Bank
status.
}
\keyword{datasets}
