% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geographies}
\alias{geographies}
\title{World Bank Geographic Entities}
\format{
A data frame with the following columns:
\describe{
\item{geography_iso3}{ISO 3166-1 alpha-3 code, where available}
\item{geography_iso2}{ISO 3166-1 alpha-2 code, where available}
\item{geography_name}{Standardized name of the geographic entity}
\item{geography_type}{Type of entity: "Country", "Region", or other
classification}
}
}
\source{
World Bank World Development Indicators API
}
\usage{
geographies
}
\description{
A dataset containing countries, regions, and other geographic entities from
the World Bank's WDI API, enriched with data from other World Bank sources.
}
\keyword{datasets}
